Handle case where task being routed has unknown config. (#330)
When the task being routed is not in the BaseCeleryConfig task_router dict the router `route_tasks_based_on_user_plan` was throwing an error `TypeError: string indices must be integers` because the fallback for `task_config` was a string instead of a dict in the format `{ "queue": "queue_name" }` (i.e. I sabotagged myself 😅). This fix routes unknown tasks directly to the default task name.

I suspect that this can happen specially with cron tasks from the worker (like the health check task or the hourly check task).