Revert "Revert (#1017) - upload finisher task is failing the dynamic routing (#1018)" (#1019)
This reverts commit fef75e68be1c90a077c6cf1ea4b83c4bb58fd296.
And re-enables the dynamic task router.... again.

This time the UploadFinisherTask, when calling ComputeComparison task was passing the arguments via `args` instead of `kwargs` as everywhere else. This breaks the expectations of the task router (that depends on `kwargs` to find the user plan).

Since the trend all over the worker is to pass arguments via `kwargs` I don't think it's a problem if we just fix this one place. If needed we can look into making the task router more robust, though.