Adds Sentry Release Tracking for Staging Env
On a staging deploy this should attempt to create a release and notify sentry. Default behavior will be to tag the release with a commit. `sentry-cli releases propose-version` should do this. 

Not thrilled about a `curl | bash` for obvious reasons. If this works, perhaps we can find a better way.