Dynamic task router (#1233)
Related ticket: [CODE-2328](https://codecovio.atlassian.net/browse/CODE-2328)

Introducing the `route_task` function to the signature process of tasks. That gives them a specific queue to be put in, that will change only for enterprise cloud users (they get a special queue because they are special). That way - using the signature method - we don't have to change the celery config defaults from shared (but notice that the choice of queue comes from shared).