Enabling compression on GCP
Implementing logic so that, if the content of something was not already compressed, we compress it to send to GCS.

Also added tests to ensure that two cases work:

- Brand new files are correctly sent to GCS
- Files that already exist in non-compressed form are properly adjusted to gzip

This seems to work well. From looking at the cassetes generated, it seems that the GCS client sends the new encoding with every upload, so the appropriate encoding is always correct.