Add support to list app deliveries in GitHub (#332)
* Add support to list app deliveries in GitHub

Related Ticket: [CODE-2489](https://codecovio.atlassian.net/browse/CODE-2489)

Add support to list app deliveries from GitHub. This change is aimed at improving reliability of gh app installs, because we plan on depending more heavily on it.

The worker will have a cron task that list webhook deliveries from time to time. It will use this new method from github. It had some unique requirements so there are some changes worth noticing.

`parse_link_header` is an utility function that parses a "link" header from GitHub. In the webhook deliveries list the next page is accessed via that link.

That means we need to return the header from `make_http_call` though, so I changed the return to include that. To make sure it's not disruptive to every current call, `api` only returns the headers if you explicitly ask for them. It's a delicate change, but won't break things (right now at least).

I decided to make `list_webhook_deliveries` an async generator function because we don't know how many webhook deliveries are there for the app, and we expect that number to increase. In any case I thought it made more sense to process the deliveries in small batches instead of accumulating everything in the function before returning, thus it's a generator.

* Creating dedicated method for paginated calls. Change return of `make_http_call` to `Response`. Remove `parse_link_header`.

`parse_link_header` was unecessary because `Response` already has `response.links` (thanks @ThiagoCodecov for the tip).

To avoid the ugly return as a tuple of the headers (or links) to `api` we're now returning a `Response` object directly. The parsing of the `Response` is moved into a separate function so it can be reused.

And finally there's a new method for paginated calls. I decided to make it a generator instead of following the GitLab implementation completely. I think there are possibly performance benefits from doing so.