Improve logs for LockError in notifications (#1044)
There are some notifications getting stuck eventually. We don't know why.
Moving forward with investigations we're trying to improve the logging of LockErrors in the notifications.

Specificly we are trying to determine if:
1. you are trying to get a lock and cannot, because some other process has it; OR
2. you’re trying to release a lock that no longer exists (i.e., it auto expires in redis, but the code thinks it’s still there)

By getting the `LockError` type we can see if it's really `LockError` or `LockNotOwnedError` (caused when trying to [release a lock not longer owned](https://redis-py.readthedocs.io/en/stable/exceptions.html#redis.exceptions.LockNotOwnedError)).

By knowing if the lock was acquired or not we can know if the issue came from within or if indeed another process had the lock before us.

Also I tested that if we exit the `with lock` block because of some random exception (say the infamous `SoftTimeLimitExceeded` exception) it will release the lock. Plus I didn't find anywhere that we manually release the lock AND the expiration time for the lock is as big as the HARD limit, so I think it's unlikely that we will see `LockNotOwnedError`, but who knows.