Removing ignore section
We previously had some files ignore section because we didn't want
test files to affect general coverage. The idea is not neither punish/reward
long/small/repetitive tests. The main thing being tracked was actual code coverage

But now we need to have that information stored for iterative testing to
be able to detect changes in tests.
We already have components ignoring the same files. So we should be able to
still see "code-only" coverage.