Adding a field "global_level_labels"
The goal is to be able to tell the customer something like "Your changes affected lines that are considered global level lines. Here are the tests that possibly depended on those global level lines"

By making this a special field, the customer can decide better whether they want to run those tests or not. I imagine this might have different effects on different use cases, and having them separate puts the control in the client