Moving carryforward overwrite logic to worker
As described in https://github.com/codecov/shared/pull/342 , we need to buble up more information about the carryforward to the worker, and it was going to take a lot of changes to do it the way it is currently done in the current `add_session` logic. So we are adding a `_adjust_sessions` function, that is going to do all the overwriting. It includes both:

- Current CFF overwrite logic (`upload` carryforward mode): If one existing report session is carryforwarded and shares at least one flag with the newly-uploaded flag, and that flag is a `upload` carryforward one (as described in the YAML), then the existing session will be fully deleted
- Iterative Testing CFF overwrite logic (`label` carryforward mode): If one existing report session is carryforwarded and shares at least one flag with the newly-uploaded flag, and that flag is a `label` carryforward one then we will check for all the labels in the newly processed report and remove all those labels from the previously mentioned report session

This change, as it stands, is a no-op. It will only apply when the shared PR gets merged (and a lot of tests will break, hopefully)