Upgrading shared
Upgrading shared to incorporate https://github.com/codecov/shared/pull/342

This makes it so the `add_session` logic now does not delete CFF sessions itself: another piece of logic will do that

So most places that use add_session will have session results changes. At least the session number, as it now picks a session number while the old session is still in place. Before, it would just pick a session number after the deletion, which would most of the times pick the deleted session number

With this, the report totals should not change compared to what it did before. So we need to watch to make sure it doesn't