Only pop lines when carryforward shift if we have it in the report (#353)
We have seen errors in which the carryforward shift gives OutOfIndexError trying to pop lines that don't exist in the report.
Apparently it's possible that we don't have all the lines in the file on the report. Particularly lines that are not code (e.g. comments) can be not-present in the uploaded reports. However if there are changes in that part of the file with removals, we get an error.

To fix it we simply verify that we can pop a line before actually doing so. Processing doesn't change, and we won't delete less lines than we would have otherwise.