Gear GitHub webook retry task to run every 6h (#1088)
We are seeing that the task responsible for requesting redelivery of webhooks is consistently timing out (every day).
It is very likely that the amount of webhooks we receive is too big for the time we can run the task.
On top of increasing the time for that task we're also gonna be running it multiple times a day (every 6h).

I reduced the time filter span so we have only a little bit of overlap between the task runs through the day.
That should be ok. And will reduce even more the number of re-deliveries a task has to do.

I also added a log at the end of a request page, so we can have intermediary logs to help us understand a little better how the task is working, and if redeliveries are being successful.