Adopt formal naming convention for task names (#359)
* Adopt formal naming convention for task names

We are interested in having special config for enterprise-cloud queues.
There is currently a suggestion to extend install YAML configuration to allow for this (see commit fa299f72b04436abcc3aa769180eb6c30165ce1e)

The issue we're facing now is that there's currently no way to know what configuration key is used by a task.
That's necessary so we know where to look for the enterprise specific config for that task (task group actually).

So the suggestion here is to use the task name to have a link to the config, through the config_group.
Notice that the changes proposed are not extensive, because it mostly formalizes a convention that has been more or less followed already. However by formalizing and depending on it we are ADOPTING A NEW CONVENTION THAT WILL HAVE TO BE FOLLOWED IN THE FUTURE.

This change should not be taken lightly.

The convention proposed is the following:
>  task_name |- app.<type>.<config_group>.<identifier>
>  <type> can be "tasks" or "cron"
>  <config_group> is the key under install config for the default queue for this task
>  <identifier> is the task name (usually same as task class)

Where "config_group" is the most important piece as it relates to enterprise queue config.

Other changes are made to update the install config and task names to reflect the adoption of this naming convention.

* Create Enum for TaskConfigGroup

Making the task naming convention stronger and more-difficult-to-escape-from we are adopting TaskConfigGroup enum.
All tasks should belong to one of the groups in the enum.
In the near future, we will be adding a `config_group` attribute to tasks that will error if not defined (to be done in the worker, where tasks live).

Doing this also allows us to maintain better parity between the install YAML and the configuration.
To do that the install yaml validation was changes to have `keysrules` that are the TaskConfigGroup's values + 'celery'.
The `valuesrules` have 2 possibilities: the old celery schema and a schema for regular tasks.

Yes, in a sense the validation is a bit more relaxed because fields particular to any task (e.g. 'timeout' for notify) can now be used by any task. BUT we don't return errors for install yaml, so technically users could do whatever they want anyway.

While it's true that we can use a function to validate the values, I thought it would be hard to read and maintain, so I opted to keep the way it is, loosing a bit on strictness but gaining in readability. In general the install yaml is not used by customers anyway (just enterprise customers).

Another change is in the config directly. You can see here the docs: https://docs.celeryq.dev/en/stable/userguide/routing.html#automatic-routing
Celery allows for pattern matching, so we can group the task config groups by route as well. I did so for any group that has more than 1 task.