Revert "Revert "Inject timeouts for tasks depending on the task routing (#1106)" (#1108)" (#1109)
This reverts commit a0fd284ca9b04d5f747d27b3b2b327f7cdddfb6f.
We had reverted that commit because errors spiked. This is due to a change in the naming of some tasks.
The spike lasted 1min before coming back into place. It was because some tasks were in the queue and got picked up by a new worker version. Obviously I panicked and rolled back, but afterwards things smoothed down and it seems to actually be working fine.

So putting this back online.