Add stack info to SessionTotalsArray __init__ warning (#368)
We have a warning in the `__init__` method of `SessionTotalsArray` to indicate that we tried to build one with data encoded in an unknown format. When deploying it for the first time we have several of those messages on the logs.

The unfortunate thing is that we don't really know where they're comming from. Happening in both api and worker make this issue harder. The working theory is that it might occur when transforming the new-encoded format in a list by accident.

Adding stack_info to the warning will help us debug.