Don't ask to activate renovate[bot] user (#1150)
[Renovate](https://docs.renovatebot.com/) is a service for
> Automated dependency updates. Multi-platform and multi-language.
Some clients use this. When the bot opens PRs they get the message "this user needs to be activated".
This means consuming a seat in Codecov for using a different service. We don't want that.

In similar fashion to what we had for dependabot, we will be tagging the bot from the email,
and also (redundantly) by it's user ID.
The redundancy is on purpose. Based on [this thread](https://github.com/orgs/community/discussions/24664#discussioncomment-3244950)
we can see that the email format might change without notice, so the user ID is an extra safety.

Notice that [from renovate docs](https://docs.renovatebot.com/getting-started/running/) there are 3 ways to run renovate[bot].
This approach only covers the 1st method (Mend Renovate app on GitHub).
The other methods are not supported by this. There is some discussion going on howto/if we will support them,
but at this time no decision has been taken.

Ticket [CODE-3283](https://codecovio.atlassian.net/browse/CODE-3283)