Don't enable docker experimental features (#1539)
[Previously](https://github.com/codecov/codecov-api/pull/1511) we removed the --squash option
from the worker enterprise build. This is because CircleCI is changing their infra
and we won't be able to use that anymore. Apparently this will be fine.

Currently I'm seeing some builds failing on an infra level because of us
activating the experimental features of Docker.
So I'm just removing that step.